+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/adapter-interfaces
 
+## 0.11.1
+
+### Patch Changes
+
+- Update the application types
+
 ## 0.11.0
 
 ### Minor Changes